add view classes for the various calendar views Rearrange the code for calendar views:
Create CalendarDay, CalendarWeek, CalendarMonth and CalendarYear classes
and associated interfaces that just hold a reference to the calendar and
time period in question.

Convert the existing DayView, WeekView, MonthView and YearView classes to
view classes associated with the above interfaces.  This way they have
access to the request object, so will be able to find the timezone of the
principal.

Remove the ICalendarDayView, ICalendarWeekView, ICalendarMonthView,
ICalendarYearView, ICalendarDayInfo, ICalendarMonthInfo and
ICalendarEventInfo interfaces.