Fix bug 43115.
The search page was using the meta_binaries call on a
sourcepackagerelease object, which makes no sense since it would return
ALL binary packages for ALL distroreleases, some duplicated.

This is fixed by moving the meta_binaries call to
distroreleasesourcepackagerelease, where it will only return binaries
pertaining to the corresponding distrorelease and sourcepackagerelease.
The search page template is updated to reflect the new call and has a new
<div> with id "binaries" to assist the new page test,
10-distribution-package-search.txt.  It also has an id "binlink" for the
binaries hrefs, so that we don't need to know the URL in the test.

A new test is also added in publishedpackage.txt to call the new method.