Convert BugMessageAddFormView to be a LaunchpadFormView.
Added an error condition if the user doesn't provide a comment nor attachment.

Changed the page template to use **launchpad_form/form.

Modified the doctests to use the new LaunchpadFormView interface.

Added a new template and class for BugMessageAddSubFormView to account
for the two ways comments and attachments can be invoked -- at the
bottom of the bug index page and through the left actions portlet.