Increase the branch mirroring frequency to every 6 hours.
Little story: The mirroring frequency was initially every 24 hours. Mark
checked out bzr from Launchpad to run some performance tests, and was suprised
to find that the branch was several revisions out of date. This is normal for
mirror branches. Since we do not yet have a way to set custom mirroring
frequencies, and after discussion with Steve, we decided to increase the
hard-coded mirroring frequency. Every 6 hours seems a good balance between
keeping the mirror branches up to date, and avoiding to hammer users servers
too much.