soyuz-set-of-uploads.txt uncovered a bug where uploads rejected before a valid PackageUpload was created caused an exception because emails are obviously sent from that class now.  In this case, a dummy PackageUpload is created since no information from it is used in the rejection email anyway.
Some changes files were also updated because email addresses were not in LP
sample data, which caused emails to not be generated.  This is only a problem
when the test depends on the debug output as some now only gets generated
when the email is sent.

I am also working around bug #117557 until it gets fixed.  See inline XXXes.