Flesh out the XMLRPC implementation of IRequestedMailingListAPI, along with updated doctests.  This no longer tests against xmlrpc.launchpad.dev because we're not sure exactly how the XMLRPC service will be exposed.  It almost definitely won't be by the public interface.
For now, the doctest tests against the view methods directly, which is enough
to ensure that it works, and is easier to debug.  There are still some XXX
comments left that will be cleaned up next.

Added an IMailingListApplication and MailingListApplication, though these may
go away before the branch is complete.