We don't need the notify_modified stuff to report the SQLObjectModifiedEvent. This isn't (apparently) how the database gets properly synced for the test suite.
Hack the internal 'direct-view' test so that when a Fault is returned from a
method, we turn it into output that matches what the external 'xmlrpc
ServerProxy' output looks like, namely, an exception.

For some unknown reason, we must use LaunchpadZopelessLayer instead of
LaunchpadFunctionalLayer to make the external test work.  But then we also
need to do some ugly post-test permission cleanup.  I need to better
understand why this is.