First crack at the implementation of Mailman's side of the XMLRPCRunner.  This puts the actual runner module in monkeypatches and the build process writes a very simple Mailman.Queue.XMLRPCRunner module.  Do it this way because then most of the code lives in our tree and the dynamically built Mailman tree has nothing that isn't generated.
buildmailman.py: Create $VAR_PREFIX/backups which is where we'll put list dir
gzip'd tarball files when we deactivate a list.  Deactivation to Mailman means
deletion, but we want to save the tarball just in case.

Added xmlrpc_url and xmlrpc_runner_sleep as parameters to configure in
launchpad.conf.  Added these to the schema, but they'll need documenting (and
other changes in this file will need to be backed out for production).

The interface calls the action 'deactivate' but the test and implementation
called them 'deactivated'.  The former is better so it is used consistently
throughout.