Implementation of private XML-RPC ports.  Here's the approach:
We define a new server type called PrivateXMLRPC and stick that on an unused
port, say 8087.  We also add a new hostname called
xmlrpc-private.launchpad.dev but this shouldn't be exposed via /etc/hosts or
proxy.  It's simply a placeholder for the internal dispatch machinery.

In systemhomes.py, we check the wsgi environment's SERVER_PORT and look to see
if it matches a configured server with a type called PrivateXMLRPC.  If so,
then we shuttle this request off to the xmlrpc-private fake internal host.
That hostname uses a private publication and request.  Ignore the publication,
it doesn't do anything special.  The request on the other hand has a special
marker interface indicating that it came from the private port.

Next, LaunchpadRootNavigation is (still) used to expose the XMLRPC endpoint,
such as /mailinglists and /bugs.  But the traverse() method. checks both the
utility and the request.  If the utility has a marker interface indicating
it's a private api, the request must have that private port marker interface.
If it does, we return the utility, otherwise None.