Mailing list subscriptions.
This branch allows people to subscribe and unsubscribe for an active mailing
list.

Added IMailingList.subscribe() and .unsubscribe() along with the
implementation of these interface methods.  Added property
IMailingList.addresses to return the text email address that are subscribed to
the mailing list.  Text addresses are returned since these are more useful for
Mailman, but it's possible I'll change this to return IEmailAddresses and
extract the text in the XMLRPC interface.

Added exceptions CannotSubscribe and CannotUnsubscribe.

Added property IMailingListSet.active_lists to return the set of all active
mailing lists.

Added class MailingListSubscription (and its interface) to track the database
table of the same name.

Added ValidMailingListVocabulary to return the list of all active mailing
lists, which may optionally match a team name (since teams and lists are
linked one-to-one).

Added and updated doctests as appropriate.