Handle mailing list membership updates when subteams are removed from their super teams.  In that case, list members with no other direct or indirect membership in the list's team are unsubscribed from the list.
Added IMailingListSubscriptionSet and MailingListSubscriptionSet to handle
this.  Essentially, it's just a utility that implements deleteSubscription()
given a person and the team they are being removed from.  Then in
TeamMembership, when the team participation is being cleaned up, call this
method to also clean up any related list membership.  deleteSubscription()
ignores cases where the team has no mailing list, and where the person is not
subscribed to that mailing list, since these are uninteresting.