Mark wants to preserve mailing list preferences (including the subscription address) when team members leave and re-join their teams.  Say Person1 is a member of Team1's mailng list List1.  Person1 leaves Team1, thus being technically unsubscribed from List1.  Later, Person1 rejoins Team1, automatically becoming resubscribed to List1.
The single outstanding problem comes about in the following situation: say
there is a Team2 which is a subteam of Team1.  Person2 joins Team2 and because
of their indirect membership in Team1, they are allowed to, and do, join
List1.  Now they leave Team2 which gets them unsubscribed from Team1.  Later
they join Team1.

Question: would they automatically get re-subscribed to List1?  Currently
that's the case, but I find that somewhat surprising.  I'm not sure the
current data model supports any other behavior though, so if the answer is
"no" then there's work to do.