Hopefully final round of reviewer comment inspired changes.
In mailinglist-subscriptions.txt, in the alternative email helper function,
Salgado suggests to use person.validatedemails.  This requires the addition of
a few flush_database_updates() but Salgado felt that was fine (these will
likely go away when we switch to Storm).

Backport from the mlist-subs-2 branch the setting of the alternative email
status to VALIDATED so that it doesn't appear as NEW and lets the above work.
I'll have to remove this from the child branch before that lands.

After discussion on IRC about the semantics of
ActiveMailingList.__contains__() when passed an object of an unexpected type,
it is deemed better to break Python's rules deliberate so that OOPS will be
generated and underlying bugs will not be masked.  This change is made, with
the additional of an explanatory comment.