Added a test for deactivation of mailing lists.  This requires a Mailman bin/withlist script called 'mmpaths' to print some important information.
Refactor 'auth' and 'HERE' to the globals created by runtests.py.  Also, catch
Exceptions and print their tracebacks.  They are unexpected failures.

XMLRPCRunner: Create the proxy in the constructor and use self._proxy
everywhere.  This appears to fix a nasty memory leak which is probably caused
by the xmlrpclib.ProxyServer (although this conjecture is totally
unsubstantiated).