Change the name of the teams and lists being used in these tests to start with 'itest-' to reduce the possibility that they will conflict with sample data.
Change the 'wait-for' polling to instead of polling for the specific data
we're looking for, to polling for Mailman having Done Something (as evidenced
by its writing to logs/xmlrpc).  Once we know its done some work, we can check
the specific condition we care about.

To make the above work, the XMLRPCRunner must always write to its log file
whenever it completes its work, even if there was nothing to do.