Clean up style of h1, h2, h3, h4, h5, and h6.
Set the padding-bottom of h1 and h4 to 0. Was 10px.

Set the margin-bottom of all to 0.5em. Was 0 for h1, 0.75em for h2,
0.5em for h3, and 0 for h4, h5, h6.

Set the margin-top of h2, h3, h4, h5, h6 (and NOT h1) to 1em. Was 1em
for h2 and h3, and 0 for h4, h5, h6.

Preserve the 0 margin-top of h1.

NOTE: h4, h5, and h6 are never used in pages where this stylesheet
applies.