Fix bug 157016 (Ugly white space between headings and tab)
Remove the additional vertical spacing introduced by the fix to bug 151429 when the structural heading image was present. Make the heading look the same whether the image is present or not.

Add some spacing below #globalsearch to provide visual separation when it wraps around.

Add lots of comments to explain non-obvious settings and interactions, so other people are less likely to break the layout of pages that use #structuralheading and #globalsearch.

Remove the obsolete (pre-1.0) "#globalsearch #q" rule.