Fix bug 157124 (minimal width is wider than necessary).
#nonhelp min-width is no longer needed since bug 151429 was fixed.

Clear the #mainarea min-width. Instead, use a (much smaller) min-width on body.

Modify the #yourlocation, #menuroot, and #logincontrol rules in various ways so the logincontrol can wrap around gracefully in narrow windows, even when the font size was changed by "zooming".

Fix the lack of space at the left of #logincontrol.

Combine the obscure "#locationbar ul" rule with the "#menuroot" rule.

Removed an obsolete "body" rule.