Implementation of team-mailing-lists-reactivate blueprint.
Added an integration test for ensuring that mailing lists can be reactivated,
and that team members will retain their preferences (i.e. subscriptions) after
reactivation.

In order to make this work, I had to change the way subscriptions are handled
when a list is deactivated.  Previously, we would clear the
MailingListSubscription table entries for all subscribers, but this breaks
preference retention.  Now, we leave these entries alone, but ensure that
getAddresses() only returns subscriptions for lists which are not INACTIVE.

Another change is made so that the archive tgz file created when a list is
deactivated has paths relative to $VAR/lists so that unpacking on
re-activation works better.

Factor out mbox iteration into itest_helper.mbox_iterator().

Added regexp filtering to the runtests.py integration test runner.