Implementation of blueprint team-mailing-lists-restrict-team-rename. Specifically, if a team has a mailing list, it cannot be renamed.
This adds two checks to guarantee a team cannot be renamed.  First, at the
database layer, setting Person.name is overridden to raise an AssertionError
when a team with a mailing list has its name set.

Second, the 'name' widget for changing a team's details is set to read-only
when the team has a mailing list.