Implementation and tests for team-mailing-lists-archiving blueprint:
https://blueprints.launchpad.net/launchpad/+spec/team-mailing-lists-archiving

This branch adds a configuration variable called archive_url_template in the
<mailman> section of launchpad.conf.  This is a string.Template style
substitution template that accepts a 'team_name' to point to the list-specific
mailing list archive.  The format is configurable because it will be one thing
for the temporary Pipermail archives, but likely something different when we
switch to a permanent archiver.

The mailing list archive is available for any team that has or has ever had an
active mailing list.  We include deactivated lists here because the policy is
to never delete archives.  The page template for the team's overview
(i.e. person-index.pt) now includes a link and url to the archive if the team
has a mailing list and that mailing list was ever active (as given by states
that can only follow MailingListStatus.ACTIVE).

This branch includes tests of the new API methods for database level Person
and MailingList classes, as well as page tests for proper display of the
archive url.

This branch also refactors mailing list review slightly, but mostly for
convenience.  More will be performed as part of bug #163130.  This branch also
cleans up some 'make lint' warnings that were reported in files I've touched,
and whitespace here and there.

One icky implementation note: IPerson.archived_mailing_list should really be
defined as an Object(schema=IMailingList) but this isn't possible due to
circular import nastiness.