Changes in response to jml's review:
The rather crufy 'archived_mailing_list' property on Person has been removed.
Person's really shouldn't contain the logic for when a mailing list could have
an archive or not.  That should properly live in the MailingList class.

Instead, IPerson has a mailing_list property which returns the IMailingList
instance for the team's mailing list, if there is one.  IMailingList now has
an archive_url property which returns the url if the mailing list has ever
been activated.

Similarly PersonView now has a cached property archive_url, which returns
either the url string, or None if this is not a team with a current or once
active mailing list.

Included here are several other smaller cosmetic changes requested by jml, and
updated tests.