Several changes to prove that using a mailing list as a contact address works, and the changes to make that work. ;)
Updated the README.txt to include instructions on adding a
+mail-configure.zcml file to talk to the itest's smtp2mbox script.

Fleshed out itest doctest to subscribe a person to the mailing list, and
submit a question, which is guaranteed to trigger a message to the project's
contact address (i.e. the mailing list).

Clean up the AnswerContact and PersonLanguage tables in the tearDown, to avoid
constraint errors.

Extend smtp2mbox to accept messages both destined to Mailman (which it will
inject into Mailman's incoming queue), and to the real smtpd
(i.e. localhost:25).

Fix the lphandler.py to pre-approve any message that has X-Generated-By header
with value 'Launchpad (canonical.com)'.  This is spoofable, but it's the best
we can do right now for accepting auto generated messages.