Changes in response to Francis's review.
Put a .reset() method on SMTPServer and actually move all the tempfile
calculations to the class itself.  While this requires us to add some
convenience methods to the class, it actually makes things much cleaner.
Plus, because we need to implement a channel subclass in smtp2mbox, it also
affords us the ability to shut the bloody broken pipe messages up.  And there
was much rejoicing.