Finish the page test for mailing list approval.
Change the form keys to use colons as delimiters instead of underscores since
the latter (but not the former) can be used in team names.

Handle None as a possible action coming back from the form (double check
this).

Include the team name in the AssertionError.

Include a link to the team page in the info notification.