Remove the temporary mlist-admin.py script and doctests.
xx-team-mailinglist-subscription.txt is moved to mailnglists/subscriptions.txt
and updated to use the new admin screen.

Because of Zope crack, use dots as the separators between form key names.
Colons are used by Zope to magically specify the type of the field.

Remove mailinglists_helper.new_list_for_team() as it's no longer used.  Add a
MailmanStub class which pretends to be the Mailman XMLRPC process.