Rework the wait_for_mailman() synchronization trick so that it's much more reliable.  Now, XMLRPCRunner will write a serial number to logs/serial every time it sees an actual change.  Previously we were syncing on writes to logs/xmlrpc which happen everytime XMLRPCRunner talks to Launchpad.
This is much more reliable, and not much more complex.  So the sync on
logs/serial knows that when it sees a modtime bump to this file, something's
changed in Mailman.  Note that while we append a monotonically increasing
serial number to this file, we really don't care about the number per se.  All
we care about is the modtime on the log file.

Another cool thing: we can reduce the LOG_GROWTH_WAIT_INTERVAL to 5 seconds
(it could probably be even less, but ditching 25 seconds is a big win anyway),
and we only need to wait for one change to the log file, so no multiple cycles
through the wait loop.

This does mean several tests now get Time outs when they didn't expect them
before, but since the wait is only 5 seconds, it doesn't hurt us much and is
easy to add.