Although unworking, this adds a test and fix for bug 208431, where reportStatus() XMLRPC calls can get lost causing lists to get unsynchronized. This adds an additional getPendingActions() action called unsynchronized which the XMLRPCRunner can use to resync the state of lists.
There are two things with this current implementation that I don't like.  I
had to add a specific permission to allow the recovery.txt test to edit
MailingList.status directly, and the flushes don't work to get the app server
to see the changes.  This makes me think that I don't have something quite
right <wink>.