Fix the librarian configuration for testrunner-appserver.  Also, update the open() urls in postings.txt. These help postings.txt pass.
When raising UploadFailed due to socket.errors, include the upload host and
port in the exception message.

harness.py: Don't put the librarian log files in the librarian root for the
test harness, otherwise they get blown away at the end of the test, and
there's useful information there.  Instead, put them in the parent, or in
tempdir.