Modify the permission for a team's +mailinglist page to allow mailing list experts, launchpad admins, and team admins to view it.  Create a new permission launchpad.MailingListManager to control this.
When requesting the team mailing list creation, remove from the view, the
check that the team does not have a mailing list.  The content object already
does this check, so it's redundant.  Worse, it's incorrect in the face of
purged lists, so why duplicate code?

For mailing list status messages, a purged list is as good as a non-existent
list.

Add Person.__repr__() sto get more information when displaying a person in an
interactive prompt.

Export a symbol 'stop' into page tests so it's easy to add a pdb breakpoint.