Fix for bug 288405, by conditionalizing (in the views) whether the +contactuser link says "Contact this user" or "Contact yourself", as per discussion with Martin.
Both the PersonView and the EmailToPersonView were extended to include a
property called specific_contact_text which check the context against the
logged in user and returns the appropriate text.  There might be opportunities
for refactoring this common code, but it also seems just as easy to code it
inline.

I also took the opportunity to turn a method into a property, and to refactor
the person-pages.txt doctest.