Preliminary step in the redesign of projects/+new according to https://dev.launchpad.net/GuidedProjectRegistration.
This simplifies (somewhat) the ProductAddView and does not make it inherit
from ProductAddVewBase (for now).  This adds a custom ProductNameWidget and
changes the button to 'Continue' for its two-step product creation.   The
schema description for 'ID' (a.k.a. 'URL') is slightly modified.  The
product-new.pt reorganization and simplification has begun.  I also removed
the +new-guided zcml, though more links need to be patched.