Fix the navigation menu.
It doesn't help for links to be in the mixin, it has to be in the subclass.  I
don't know why and right now I don't care.

TopLevelContextMenuMixin can't be a ContextMenu because it's used on some old
pages and new pages.  The latter must mixin NavigationMenu, and the two are
mutually exclusive.

Add the tal magic to throw the nav menu on the /people page.