Update /projects to UI 3.0.
Add a ProductSetNavigationMenu.

Refactor RegistryCollectionActionMenuBase so that `links` is not defined here,
but on the concreted base class.  The menu used on /people has merge actions
which aren't relevant for /projects.

Update the template and put the stats under the heading, in the same location
and style that it's in for /people.