Rework the version.txt extraction to be more robust.
Add IEditableContextTitle to mark views that have an editable title.

WatermarkTalesAdapter adapts views now, not contexts.  That's because the view
needs to be available, which is fine because the context is available from the
view.

Fix the implementation for views which provide IEditableContextTitle.

Code cleanup.

Make IProduct's index view an IEditableContextTitle.