Convert ~person/codesofconduct pages to UI 3.0
Along the way, move the page title for ../codeofconduct/1.0.1 out of
pagetitles.py, and simplify CodeOfConductView.

Also, get rid of the crufty triggering of the "changes to code-of-conduct
signature records" in PersonCodeOfConductEditView.  This used to be triggered
by a TAL directive, but that's insane; it is now done in the view's
initialize() method.