In utilities/community-contributions.py list Данило Шеган (Danilo Šegan) in Cyrillic too, in the 'known_canonical_devs' list.
This is also important because there's a bug in either editmoin or
urllib or both that causes us to get a "KeyError: u'\u0434'" traceback
if Danilo shows up as a contributor.  At some point, this is going to
bite us with someone else, and we'll have to really fix it.  But
listing Danilo in Cyrillic is necessary anyway, independently of that.