In PropertyFile.parse(), there is an XXX from Danilo which says: "...This call unescapes everything so we don't need to care about  quotes escaping."
Well, this was true in Python 2.4, but only by a mistake which has since been
corrected in Python 2.5.  This means that unicode-escape codec does not by
default "unescape everything".  raw-unicode-escape will do that, and since
the intent of the code is to have "everything escaped", we use that to create
the 8-bit string from the unicode.  Then unicode-escape to go from the 8-bit
string back to the unicode will do the right thing.