update-sourcecode: Avoid pulling branches which are already at the correct revision.
Prior to this, update-sourcecode unconditionally 'pull --overwrite's each
branch. This causes bzr-dbus/bzr-gtk notifications to be issued for every
sourcecode branch, even the unchanged ones, which spams notification-daemon or
notify-osd quite unattractively.

Also, since this implied a change in the console output anyway, fix it so that
the bzr progress bar does not overlay and remove the messages saying which
branch is being pulled.