Make ICodeImportSet.new act on IBranchTargets.
The code import creation only cares about a namespace, so use the more
generic interface to get one.

This changes the factory, adding makeProductCodeImport to still allow you
to pass product which will be turned in to an IBranch target as needed.

There is also makeAnyCodeImport for the test to declare that it doesn't
really care what the target is.

This doesn't add any new functionality, that will come later.