Remove unneeded visible_comments_from_display member
Since this routine has to create a copy of the list of comments provided
by visible_[oldest|newest]_comments_for_display, that really diminishes
it's usefulness in general, to being a convenience routine for test
code.  But why not just update the tests to run against the underlying
routines directly?