Refactor code to use an activities dict instead of list
The activities are initially already in a dict, so we simplify things by
just making a copy of that, removing activities at the point they're
added to the comments, and then converting the remainder into a list at
the end.  This simplifies the activity retrieval code enough that we
don't need it broken out into a separate routine, so we can drop
listActivityForComment().

In addition, since this refactoring eliminates a couple list iterations
in favor of dict lookups, this should result in a net performance
benefit, although this will probably be unnoticeable except maybe on
bugs with a *lot* of comments (if they don't time out anyway).