Fix some fallout from the changes.
lib/canonical/launchpad/doc/tales.txt was asserting something about
the state of a Factory object without passing it as an argument, so it
failed when I had the Factory use a different default from the object.

lp/blueprints/stories/standalone/sprint-links.txt was broken because it
was asserting that "Proposed" isn't in the page, but it is now due to
the webservice/javascript stuff, which puts that string in there, as
part of a <script>, for another attribute. Just dropping the check
makes sense, as it's hard to avoid the issue, and the templates don't even
work the same way any more, so the test wasn't useful anyway.