add some tracing, fix some bugs
I was using the file object, which select is okay with, but
it was returning the file-id itself, and we were then
mapping using the file and not the file-id, which was bad.

I'm not getting a POLLHUP for the fifos, so I'm not really
sure how to know when the remote service has shut down.
It won't matter in practice, because we won't be using a
poller, like we're doing here (twisted will be doing that).

I've been trying to do some timing tests. As near as
I can tell, after connect_to_* has started up, it
is able to connect, send the request, and get the hello
response in about 80ms. This is compared to 900ms to
run 'bzr serve --inet', and 2400ms for lp-serve.