Start building the testing infrastructure.
Move lpserve into a plugin module rather than a simple script. The main
benefit is that I can now invoke the test suite directly, rather than
having to use the launchpad test runner, which is much heavier weight.
(It takes >10s just to load the tests to filter it down to the test I
want to run, I can run all tests under bzr selftest is 1.4s)

Create a subclass for testing LPServe in the test suite, that makes
it a bit easier to make sure that the class is started and stopped
appropriately.