Start trying to get some subprocess style testing for the service as well.
I'm running into problems because of path issues. Specifically I run with a relative-to-cwd
path to 'bzr' and TestCaseWithTransport does a chdir, but sys.argv seems to show only
the relative path. So it cannot find the subprocess.
Running with an absolute path to bzr fixes that problem, but I have the feeling
it will confuse the zope test runner. So I'll need a different fix for that.
After working around that, the next failure is because the BZR_PLUGIN_PATH isn't
set correctly. I don't have any quick answers here yet.