Do lots of dancing to get it to work.
It looks like ultimately we'll be back to running only in the zope test runner.
The issue is that you have to use lots of 'canonical.' and 'lp.' code to determine
where the 'bzr' and the 'python' scripts you need are. Which is duplicated with
the code in 'codehosting/test_lpserve.py'.

But for now, I'll live with this, because the tests are still faster.
The one fork test really is slow, at about 4.3s to run.

However, it already caught a bug that I was re-opening stdout for stderr,
and never opening the stderr fifo. So it did catch something important.