We are getting an odd return code.
I think this is actually because calling sys.exit() is raising an
exception, which is getting caught and adjusting the return code.

We can do a new exception instead, which I think will work.
However, we also want a good way of communicating this info
with the twisted process, and I think talking on the existing
socket is the best method, so refactoring is in order.