Upping the socket.timeout number caused some small issues (I got an erroneous failure).
However, it definitely shows up in the performance time, since the 1s delay is very obvious.
Moving the message-to-the-client into the spawned child seems to avoid the
earlier error. (My guess is it reduces the race from fork())


How about some timings:
  time 'echo hello | ssh localhost bzr ...'
  2.50s baseline codehosting time
  0.8s  time using real bzr and openssh localhost
  0.6s  time using openssh localhost and /bin/false
  1.35s time using Forking w/ 1.0s timeout
  0.40s time using Forking w/ 0.1s timeout

So we really should trap for SIGCHILD, setting the timeout
to something low is wasteful, since we end up scanning far
too often.