Create a SIGCHLD handler.
This speeds up the test suite considerably, it seems the 1.0s timeout really was
to blame. This allows us to report back much faster when a child exits.
We have to call _poll_children directly, as it appears that the signal doesn't
actually interrupt .accept(). (My guess is that it does, but python hides that
fact.)


When called sequentially, we are now down to 250ms per connection, which is
pretty darn good. We may just want to increase the preloaded modules.

However, when using simultaneous connections, we seem to be getting a lot
of random failures, so the code isn't 100% robust yet.