Include a few more modules to be preloaded.
It turns out that we did, indeed, get EINTR we were just logging it and then
ignoring it, rather than _poll_children. Now we always poll after looping.

At this point, we're starting to get some random failures, stuff like
EBADF. But it is all pretty random. async can be such a pain... :)