Don't open sys.stdin in O_NONBLOCK.
It turns out that os.open() sets the read content to be non-blocking as well as the
initial opening. We sort of wanted to avoid blocking on the open() call, but
we want the read stream itself to be blocking.
Anyway, this seems to work, so we'll go with it.