Untangle circular class definition issue
BugTracker refers to BugTrackerComponentGroups, but the latter also
contains a link to the BugTracker.  We need to delay definition of the
schema for the references until after both classes are defined.  Since
most users will need BugTracker but not BugTrackerComponentGroup, allow
the latter to be the late definition, as it can remain undefined if it's
not needed.