Actually, if we just use getAllRemoteComponentGroups() we can drop component_groups entirely.
While it had seemed better to use component_groups, it seems that a
SQLObject container object with Storm-based components leads to problems
that I'm not sure how to work around.  By using an accessor to create a
pure Storm container to iterate over the Storm objects, it seems to work
better.