Drop the stubbed in --purge-first option.
I don't expect we'd often(ever?) need to purge the data.  If it ever
happens that we do, it's probably just as easy to issue DELETEs to the
component and component group tables, since in either case this has to
be done by a LOSA.

If it turns out this functionality actually is needed, it can be added
at that point.