Add a 'bzr launchpad-forking-service' call as part of the acceptance tests SSHServerLayer.
This launches the forking service in the setUp() and should be shutting it down in tearDown().
The acceptance tests now pass, though I'll probably want some review about whether my class structure
is reasonable.

The biggest thing I don't like is that it uses the config entry for the socket to use. However, I'm
not sure how to pass that info to the newly running twisted daemon.

With this patch, all the 'test_acceptance' tests pass.