Improve SQL efficiency for searching for bugs by tag (Pre-filter by Bug.id)
Addresses LP Bug #501945:
Product:+bugs timeout with search for bugs without tags (tag:-*)

The sub-query for tags was not pre-filtered by bug ID, so was causing
a large (and unnecessary) load on the database for some queries, such
as wildcard matches.

A significant speedup is obtained on large data-sets by rephrasing the
SQL sub-query to consider only those BugTag rows which match the current
bug being considered in the outer query.