Change how the children connect to their fifos.
This code uses open() in non-blocking mode so that we can timeout and return.
This fits what I wanted, that we could wait a period of time before deciding
that nobody is going to talk to us after all.
This happens when the Conch service runs out of file handles, asks us
to start a new process, but then is unable to actually connect.

Note that current testing shows it might not be possible to change a
file handle back to blocking mode once it has been opened, because
I saw some test failures that indicated the Smart server was getting
EAGAIN, which you aren't supposed to get in blocking mode.