Track failure-to-open-handles more closely.
When spawning a new child, watch out for failures while opening, and immediately
close the other handles when stuff fails. This seems to cleanup file handles
in a much cleaner fashion.
Note that we might still have a socket laying around that is waiting for
the master process to tell us the child exited.
(If we get far enough along the process, but not far enough.)
However, with the other code for children to commit suicide after
two minutes if they haven't been connected to, that will close those
handles as well.
At least, that is what I see in testing.