Document what is going on in the child if it gets interrupted.
We can't change the test suite to not catch SIGALRM, so either we always catch
it, or we just do less when it only triggers EINTR.
This also moves away a race condition in the 'child_exits_eventually' test.
Specifically, the parent cleans up after the child after SIGALRM goes off.
However, it was reporting the child gone before cleaning up the dir, and the
test was waiting on the report to go detect it had been cleaned up.
Cleaning first means the test always sees it after the cleanup has happened.