Add synthetic bugtracker scrapers
The update-bugzilla-remote-components.py cron job can fail when
getRemoteProductsAndComponents() hits various HTTPErrors.

This patch modifies getRemoteProductsAndComponents() so that instead of
creating a new scraper object per bugtracker, to instead allow the test
code to pass in its own scraper, which is used for all requests instead
of generating a scraper object for each URL.

The test code can then supply scrapers subclassed from
BugzillaRemoteComponentScraper with hardcoded behaviors:

StaticTextBugzillaRemoteComponentScraper replaces the static text
parameter.  Instead of hardcoding a check for this option, we allow the
test code to provide the static text via the subclasses' getPage().

FaultyBugzillaRemoteComponentScraper provides a way to simulate hitting
various error codes by raising whatever errors we pass to it.

A assertGetRemoteProductsAndComponentsDoesNotAssert() helper method is
added to allow us to run getRemoteProductsAndComponents() and check that
no assertions were thrown.