Accept as trusted mail signed by the Sender domain not the From address (bug 643223).
This should allow people who use eg gmail (which signs from the Sender) but
with their own domain in the From address to use Launchpad, as long as they
have validated the Sender address.

This also adds a process-one-mail.py script that reads a mail from stdin 
and can more easily be used to locally test incoming mail code.