Add support for multiple virtualized buildd guests per host system.
The buildmaster uses a resume command that is typically ssh to the host
system with a key that triggers a guest reset.  In the old system where
virtualized guests consumed an entire host, this worked well. 

To support multiple builder guests on a single host, we must pass in the
name of the guest to be reset, so that the resume command can forward this
information further down the chain.