Tweak the test a bit.
Instead of using select() to give us a suspend, use threading events.
And then we can use 2 events to kick the thread, and wait for it
to at least wake up, before we exit the test.
This makes it very likely that we will never leak the 'wait_for' thread.