Reduce wasted memory in loggerhead.
We create a new Branch (& Repository, etc) for every loggerhead request.
The objects don't go cleanly via gc. So help them along and flush
their caches when we are done with them.

This should not impact response times, because the Branch/Repository
objects are not getting reused anyway.
In local testing, this *dramatically* reduced peak memory while
recursing over the Launchpad history in loggerhead. (like 500+MB down
to a stable 130-200MB).