Mitigate librarian swift FD leaks (bug #1408557) by reusing connections on 404.
python-swiftclient doesn't allow us to explicitly close the connections
that it makes, so it relies on destruction to clean up the connections.
But swiftclient's HTTPConnections are involved in cycles, so if GC isn't
invoked frequently the connections can pile up and result in EMFILE.

It's also possibly exacerbated by
http://twistedmatrix.com/trac/ticket/3853, but that's not the whole
issue.